Extract merging of equivalents
